requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-chess
gunicorn
